+++ tests/FSharp.Data.DesignTime.Tests/TypeProviderInstantiation.fs
@@ -28,7 +28,8 @@ type internal CsvProviderArgs =
       EmbeddedResource : string
       PreferDateOnly : bool
       StrictBooleans : bool
-      UseOriginalNames : bool }
+      UseOriginalNames : bool
+      PreferFloats : bool }
 
 type internal XmlProviderArgs =
     { Sample : string
@@ -43,7 +44,8 @@ type internal XmlProviderArgs =
       InferenceMode: InferenceMode
       PreferDateOnly : bool
       DtdProcessing : string
-      UseOriginalNames : bool }
+      UseOriginalNames : bool
+      PreferOptionals : bool }
 
 type internal JsonProviderArgs =
     { Sample : string
@@ -59,7 +61,8 @@ type internal JsonProviderArgs =
       Schema: string
       PreferDateOnly : bool
       UseOriginalNames : bool
-      OmitNullFields : bool }
+      OmitNullFields : bool
+      PreferOptionals : bool }
 
 type internal HtmlProviderArgs =
     { Sample : string
@@ -118,7 +121,8 @@ type internal TypeProviderInstantiation =
                    box x.EmbeddedResource
                    box x.PreferDateOnly
                    box x.StrictBooleans
-                   box x.UseOriginalNames |]
+                   box x.UseOriginalNames
+                   box x.PreferFloats |]
             | Xml x ->
                 (fun cfg -> new XmlProvider(cfg) :> TypeProviderForNamespaces),
                 [| box x.Sample
@@ -133,7 +137,8 @@ type internal TypeProviderInstantiation =
                    box x.InferenceMode
                    box x.PreferDateOnly
                    box x.DtdProcessing
-                   box x.UseOriginalNames |] 
+                   box x.UseOriginalNames
+                   box x.PreferOptionals |]
             | Json x -> 
                 (fun cfg -> new JsonProvider(cfg) :> TypeProviderForNamespaces),
                 [| box x.Sample
@@ -149,7 +154,8 @@ type internal TypeProviderInstantiation =
                    box x.Schema
                    box x.PreferDateOnly
                    box x.UseOriginalNames
-                   box x.OmitNullFields |]
+                   box x.OmitNullFields
+                   box x.PreferOptionals |]
             | Html x ->
                 (fun cfg -> new HtmlProvider(cfg) :> TypeProviderForNamespaces),
                 [| box x.Sample
@@ -275,7 +281,8 @@ type internal TypeProviderInstantiation =
                   EmbeddedResource = ""
                   PreferDateOnly = false
                   StrictBooleans = false
-                  UseOriginalNames = false }
+                  UseOriginalNames = false
+                  PreferFloats = false }
         | "Xml" ->
             Xml { Sample = args.[1]
                   SampleIsList = args.[2] |> bool.Parse
@@ -289,7 +296,8 @@ type internal TypeProviderInstantiation =
                   InferenceMode = args.[7] |> InferenceMode.Parse
                   PreferDateOnly = false
                   DtdProcessing = "Ignore"
-                  UseOriginalNames = false }
+                  UseOriginalNames = false
+                  PreferOptionals = true }
         | "Json" ->
             // Handle special case for Schema.json tests where some fields might be empty
             if args.Length > 5 && not (String.IsNullOrEmpty(args.[5])) then
@@ -306,7 +314,8 @@ type internal TypeProviderInstantiation =
                        Schema = if args.Length > 8 then args.[8] else ""
                        PreferDateOnly = false
                        UseOriginalNames = false
-                       OmitNullFields = false }
+                       OmitNullFields = false
+                       PreferOptionals = true }
             else
                 // This is for schema-based tests in the format "Json,,,,,true,false,BackwardCompatible,SimpleSchema.json"
                 Json { Sample = args.[1]
@@ -322,7 +331,8 @@ type internal TypeProviderInstantiation =
                        Schema = if args.Length > 8 then args.[8] else ""
                        PreferDateOnly = false
                        UseOriginalNames = false
-                       OmitNullFields = false }
+                       OmitNullFields = false
+                       PreferOptionals = true }
         | "Html" ->
             Html { Sample = args.[1]
                    PreferOptionals = args.[2] |> bool.Parse